+++ frontend/assets/shared/components/CheckboxInput.js
