+++ frontend/assets/shared/components/FloatInput.js
