+++ frontend/assets/shared/components/IntegerInput.js
