+++ frontend/assets/shared/components/QuillTextInput.js
