+++ frontend/assets/shared/components/ReactDatePicker.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import moment from "moment";
 import DatePicker from "react-datepicker";
 
-import "../.../../../../../node_modules/react-datepicker/dist/react-datepicker.css";
+import "../../../node_modules/react-datepicker/dist/react-datepicker.css";
 
 class ReactDatePicker extends Component {
     constructor(props) {
