+++ frontend/assets/shared/components/VisualToolbar.js
